car-company-seg-salary-prediction/requirements.txt
azureml_automl_core==1.52.0.post1
azureml_core==1.52.0
azureml_dataprep_native==38.0.0
azureml_pipeline_core==1.52.0
azureml_pipeline_steps==1.52.0
azureml_telemetry==1.52.0
azureml_train_automl_client==1.52.0
azureml_train_core==1.52.0
azureml_train_restclients_hyperdrive==1.52.0
joblib==1.1.0
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.2
Requests==2.31.0
scikit_learn==1.1.1
seaborn==0.11.2
